tests/krate/publish: Extract `validation` module
