+++ src/tests/krate/publish/validation.rs
@@ -1,0 +1,78 @@
+use crate::builders::PublishBuilder;
+use crate::util::{RequestHelper, TestApp};
+use crates_io::controllers::krate::publish::missing_metadata_error_message;
+use crates_io::models::krate::MAX_NAME_LENGTH;
+use http::StatusCode;
+
+#[test]
+fn invalid_names() {
+    let (app, _, _, token) = TestApp::full().with_token();
+
+    let bad_name = |name: &str, error_message: &str| {
+        let crate_to_publish = PublishBuilder::new(name, "1.0.0");
+        let response = token.publish_crate(crate_to_publish);
+        assert_eq!(response.status(), StatusCode::OK);
+
+        let json = response.into_json();
+        let json = json.as_object().unwrap();
+        let errors = json.get("errors").unwrap().as_array().unwrap();
+        let first_error = errors.first().unwrap().as_object().unwrap();
+        let detail = first_error.get("detail").unwrap().as_str().unwrap();
+        assert!(detail.contains(error_message), "{detail:?}");
+    };
+
+    let error_message = "expected a valid crate name";
+    bad_name("", error_message);
+    bad_name("foo bar", error_message);
+    bad_name(&"a".repeat(MAX_NAME_LENGTH + 1), error_message);
+    bad_name("snow☃", error_message);
+    bad_name("áccênts", error_message);
+
+    let error_message = "cannot upload a crate with a reserved name";
+    bad_name("std", error_message);
+    bad_name("STD", error_message);
+    bad_name("compiler-rt", error_message);
+    bad_name("compiler_rt", error_message);
+    bad_name("coMpiLer_Rt", error_message);
+
+    assert!(app.stored_files().is_empty());
+}
+
+#[test]
+fn license_and_description_required() {
+    let (app, _, _, token) = TestApp::full().with_token();
+
+    let crate_to_publish = PublishBuilder::new("foo_metadata", "1.1.0")
+        .unset_license()
+        .unset_description();
+
+    let response = token.publish_crate(crate_to_publish);
+    assert_eq!(response.status(), StatusCode::OK);
+    assert_eq!(
+        response.into_json(),
+        json!({ "errors": [{ "detail": missing_metadata_error_message(&["description", "license"]) }] })
+    );
+
+    let crate_to_publish = PublishBuilder::new("foo_metadata", "1.1.0").unset_description();
+
+    let response = token.publish_crate(crate_to_publish);
+    assert_eq!(response.status(), StatusCode::OK);
+    assert_eq!(
+        response.into_json(),
+        json!({ "errors": [{ "detail": missing_metadata_error_message(&["description"]) }] })
+    );
+
+    let crate_to_publish = PublishBuilder::new("foo_metadata", "1.1.0")
+        .unset_license()
+        .license_file("foo")
+        .unset_description();
+
+    let response = token.publish_crate(crate_to_publish);
+    assert_eq!(response.status(), StatusCode::OK);
+    assert_eq!(
+        response.into_json(),
+        json!({ "errors": [{ "detail": missing_metadata_error_message(&["description"]) }] })
+    );
+
+    assert!(app.stored_files().is_empty());
+}
